refactor: rename instance-list component folder
