+++ gravitee-apim-console-webui/src/management/instances/instance-list/instance-list.component.scss
