+++ gravitee-apim-console-webui/src/management/instances/instances.module.ts
@@ -19,7 +19,7 @@ import { UIRouterModule } from '@uirouter/angular';
 import { NgModule } from '@angular/core';
 
 import { InstanceDetailsModule } from './instance-details/instance-details.module';
-import { InstanceListModule } from './list/instance-list.module';
+import { InstanceListModule } from './instance-list/instance-list.module';
 import { states } from './instances.route';
 
 @NgModule({
